If JavaScript is disbled, don't underline the FAQ links. It's confusing.
I browse with JS disabled by default, which meant that reading the FAQ
presented me with what looked to be dead links. In fact, these elements
were really just activation points for a JavaScript animation. I found
that to be somewhat confusing sine, again, JS is disabled by default.

This patch fixes the issue by removing the visual indication of a link
on anchors that have no `href` attribute, but retaining the same
behavior for visitors that have JavaScript enabled.